Clientes/requirements.txt

geopandas==0.14.4

